requirements.txt
PySide2
python-vlc